Commit before switching to master
